pyproject.toml
[project]
name = "fashion-mnist-classifier"
version = "0.1.0"
description = "Fashion-MNIST classification with PyTorch Lightning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.1.0",
    "hydra-core>=1.3.0",
    "mlflow>=2.10.0",
    "dvc[s3]>=3.0.0",
    "fire>=0.5.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pillow>=10.0.0",
    "boto3>=1.28.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "ruff>=0.1.9",
    "pytest>=7.4.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["fashion_mnist_classifier", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip strict (Python 3.10+)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports в __init__

[tool.ruff.lint.isort]
known-first-party = ["fashion_mnist_classifier"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
